refactor(headless-client): remove pointless layering
