+++ rust/headless-client/Cargo.toml
@@ -15,7 +15,6 @@ connlib-client-shared = { workspace = true }
 connlib-shared = { workspace = true }
 firezone-bin-shared = { workspace = true }
 futures = "0.3.30"
-git-version = "0.3.9"
 humantime = "2.1"
 ip_network = { version = "0.4", default-features = false }
 phoenix-channel = { workspace = true }
