+++ rust/headless-client/src/dns_control/linux.rs
@@ -1,13 +1,13 @@
 use super::DnsController;
 use anyhow::{bail, Context as _, Result};
-use firezone_bin_shared::{DnsControlMethod, TunDeviceManager};
+use firezone_bin_shared::{platform::DnsControlMethod, TunDeviceManager};
 use std::{net::IpAddr, process::Command, str::FromStr};
 
 mod etc_resolv_conf;
 
 impl DnsController {
     #[allow(clippy::unnecessary_wraps)]
-    pub(crate) fn deactivate(&mut self) -> Result<()> {
+    pub fn deactivate(&mut self) -> Result<()> {
         tracing::debug!("Deactivating DNS control...");
         if let DnsControlMethod::EtcResolvConf = self.dns_control_method {
             // TODO: Check that nobody else modified the file while we were running.
@@ -22,7 +22,7 @@ impl DnsController {
     /// it would be bad if this was called from 2 threads at once.
     ///
     /// Cancel safety: Try not to cancel this.
-    pub(crate) async fn set_dns(&mut self, dns_config: Vec<IpAddr>) -> Result<()> {
+    pub async fn set_dns(&mut self, dns_config: Vec<IpAddr>) -> Result<()> {
         match self.dns_control_method {
             DnsControlMethod::Disabled => Ok(()),
             DnsControlMethod::EtcResolvConf => {
@@ -38,7 +38,7 @@ impl DnsController {
     /// Flush systemd-resolved's system-wide DNS cache
     ///
     /// Does nothing if we're using other DNS control methods or none at all
-    pub(crate) fn flush(&self) -> Result<()> {
+    pub fn flush(&self) -> Result<()> {
         // Flushing is only implemented for systemd-resolved
         if matches!(self.dns_control_method, DnsControlMethod::SystemdResolved) {
             tracing::debug!("Flushing systemd-resolved DNS cache...");
