+++ rust/headless-client/src/lib.rs
@@ -11,7 +11,7 @@
 use anyhow::{Context as _, Result};
 use connlib_client_shared::{Callbacks, Error as ConnlibError};
 use connlib_shared::callbacks;
-use firezone_bin_shared::DnsControlMethod;
+use firezone_bin_shared::platform::DnsControlMethod;
 use std::{
     net::{IpAddr, Ipv4Addr, Ipv6Addr},
     path::PathBuf,
@@ -20,75 +20,47 @@ use tokio::sync::mpsc;
 use tracing::subscriber::set_global_default;
 use tracing_subscriber::{fmt, layer::SubscriberExt as _, EnvFilter, Layer as _, Registry};
 
-use platform::default_token_path;
-use platform::tcp_socket_factory;
-use platform::udp_socket_factory;
-
 /// Generate a persistent device ID, stores it to disk, and reads it back.
-pub(crate) mod device_id;
+pub mod device_id;
 // Pub because the GUI reads the system resolvers
 pub mod dns_control;
 mod ipc_service;
 pub mod known_dirs;
-mod signals;
-mod standalone;
+// TODO: Move to `bin-shared`?
+pub mod signals;
 pub mod uptime;
 
-#[cfg(target_os = "linux")]
-#[path = "linux.rs"]
-pub mod platform;
-
-#[cfg(target_os = "windows")]
-#[path = "windows.rs"]
-pub mod platform;
-
 pub use ipc_service::{ipc, run_only_ipc_service, ClientMsg as IpcClientMsg};
-pub use standalone::run_only_headless_client;
 
-use dns_control::DnsController;
+pub use dns_control::DnsController;
 use ip_network::{Ipv4Network, Ipv6Network};
 
 /// Only used on Linux
 pub const FIREZONE_GROUP: &str = "firezone-client";
 
-/// Output of `git describe` at compile time
-/// e.g. `1.0.0-pre.4-20-ged5437c88-modified` where:
-///
-/// * `1.0.0-pre.4` is the most recent ancestor tag
-/// * `20` is the number of commits since then
-/// * `g` doesn't mean anything
-/// * `ed5437c88` is the Git commit hash
-/// * `-modified` is present if the working dir has any changes from that commit number
-pub(crate) const GIT_VERSION: &str = git_version::git_version!(
-    args = ["--always", "--dirty=-modified", "--tags"],
-    fallback = "unknown"
-);
-
-const TOKEN_ENV_KEY: &str = "FIREZONE_TOKEN";
-
 /// CLI args common to both the IPC service and the headless Client
 #[derive(clap::Parser)]
-struct CliCommon {
+pub struct CliCommon {
     #[cfg(target_os = "linux")]
     #[arg(long, env = "FIREZONE_DNS_CONTROL", default_value = "systemd-resolved")]
-    dns_control: DnsControlMethod,
+    pub dns_control: DnsControlMethod,
 
     #[cfg(target_os = "windows")]
     #[arg(long, env = "FIREZONE_DNS_CONTROL", default_value = "nrpt")]
-    dns_control: DnsControlMethod,
+    pub dns_control: DnsControlMethod,
 
     /// File logging directory. Should be a path that's writeable by the current user.
     #[arg(short, long, env = "LOG_DIR")]
-    log_dir: Option<PathBuf>,
+    pub log_dir: Option<PathBuf>,
 
     /// Maximum length of time to retry connecting to the portal if we're having internet issues or
     /// it's down. Accepts human times. e.g. "5m" or "1h" or "30d".
     #[arg(short, long, env = "MAX_PARTITION_TIME")]
-    max_partition_time: Option<humantime::Duration>,
+    pub max_partition_time: Option<humantime::Duration>,
 }
 
 /// Messages we get from connlib, including ones that aren't sent to IPC clients
-enum InternalServerMsg {
+pub enum InternalServerMsg {
     Ipc(IpcServerMsg),
     OnSetInterfaceConfig {
         ipv4: Ipv4Addr,
@@ -118,8 +90,8 @@ pub enum IpcServerMsg {
 }
 
 #[derive(Clone)]
-struct CallbackHandler {
-    cb_tx: mpsc::Sender<InternalServerMsg>,
+pub struct CallbackHandler {
+    pub cb_tx: mpsc::Sender<InternalServerMsg>,
 }
 
 impl Callbacks for CallbackHandler {
